+++ HPXMLtoOpenStudio/resources/constants.rb
@@ -9,7 +9,6 @@ module Constants
 
   # Object types
   ObjectTypeAirSourceHeatPump = 'air source heat pump'
-  ObjectTypeBEVDischargeProgram = 'bev discharge program'
   ObjectTypeBattery = 'battery'
   ObjectTypeBatteryLossesAdjustment = 'battery losses adjustment'
   ObjectTypeBoiler = 'boiler'
@@ -84,6 +83,8 @@ module Constants
   ObjectTypeUnitHeater = 'unit heater'
   ObjectTypeUnmetHoursProgram = 'unmet hours program'
   ObjectTypeVehicle = 'vehicle'
+  ObjectTypeVehicleDischargeScheduleSensor = 'vehicle discharge schedule sensor'
+  ObjectTypeVehicleUnmetHoursProgram = 'vehicle unmet hours program'
   ObjectTypeWaterHeater = 'water heater'
   ObjectTypeWaterHeaterSetpoint = 'water heater setpoint'
   ObjectTypeWaterHeaterAdjustment = 'water heater energy adjustment'