+++ Changelog.md
@@ -23,6 +23,7 @@ __New Features__
 - For storm windows, removes minimum base window U-factor limit and throws a warning instead if the base window U-factor is below 0.3.
 - Whole SFA/MF buildings:
   - Allows modeling inter-unit heat transfer using the `@sameas` attribute.
+  - Allows modeling detailed electric vehicles.
   - Documents a workaround for modeling common spaces (conditioned or unconditioned).
   - See the [documentation](https://openstudio-hpxml.readthedocs.io/en/latest/workflow_inputs.html#whole-sfa-mf-buildings) for more information.
 
@@ -36,6 +37,7 @@ __Bugfixes__
 - Fixes requested EnergyPlus timeseries output variables/meters not displayed in DView if they don't have units.
 - Fixes possible errors when small water flow rates for variable-speed experimental ground-source heat pump model.
 - Fixes possible ground-source heat pump sizing error if the heating or cooling design temperature differences are zero.
+- Fixes EMS discharge power program and assignment of default discharging schedule for detailed electric vehicles.
 
 ## OpenStudio-HPXML v1.10.0